+++ src/App.js
@@ -1,13 +1,7 @@
 import "./App.css";
 import { useState } from "react";
-import { Link, Routes, Route } from "react-router-dom";
-import Home from "./home/home";
 import { NavLinks } from "./components/navlinks";
-import Tests from "./tests/tests";
-import Recipes from "./recipes/recipes";
-import AllRecipes from "./recipes/allRecipes";
-import Navbar from "./components/navbar";
-import SideNav from "./components/sideNav";
+import { Navbar, SideNav, NavigationRoutes } from "./components/navbar";
 
 function App() {
   const [openSideNav, setOpenSideNav] = useState(false);
@@ -26,12 +20,7 @@ function App() {
           setOpenSideNav(!openSideNav);
         }}
       />
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/recipes" element={<Recipes />} />
-        <Route path="/tests" element={<Tests />} />
-        <Route path="/all-recipes" element={<AllRecipes />} />
-      </Routes>
+      <NavigationRoutes />
     </section>
   );
 }
