+++ src/components/navbar.js
@@ -1,7 +1,22 @@
 import { Link } from "react-router-dom";
+import { Routes, Route } from "react-router-dom";
+import Home from "../home/home";
+import Recipes from "../recipes/recipes";
+import AllRecipes from "../recipes/allRecipes";
+import Tests from "../tests/tests";
 
+export const NavigationRoutes = () => {
+  return (
+    <Routes>
+      <Route path="/" element={<Home />} />
+      <Route path="/recipes" element={<Recipes />} />
+      <Route path="/tests" element={<Tests />} />
+      <Route path="/all-recipes" element={<AllRecipes />} />
+    </Routes>
+  );
+};
 
-const Navbar = (props) => {
+export const Navbar = (props) => {
   return (
     <nav className="home-nav-bar">
       <div className="logo">
@@ -26,17 +41,29 @@ const Navbar = (props) => {
         </li>
       </ul>
       <div className="menu-bar">
-        <button
-          className="menu-btn"
-          onClick={props.clickFunction}
-        >
+        <button className="menu-btn" onClick={props.clickFunction}>
           <i class="fa-solid fa-bars"></i>
         </button>
       </div>
-     
     </nav>
-    
   );
 };
 
-export default Navbar;
+export const SideNav = (props) => {
+  return (
+    <div
+      className="sidenav"
+      style={props.toggle ? { left: "0%" } : { left: "100%" }}
+    >
+      <div className="links-list">
+        {props.links.map((item, index) => {
+          return (
+            <Link to={item.url} onClick={props.clickFunction}>
+              {item.name}
+            </Link>
+          );
+        })}
+      </div>
+    </div>
+  );
+};
