--- src/components/recipes.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const RecipesList = () => {
-    return (
-        <div>RecipesList</div>  
-    )
-}
-
-export default RecipesList
