+++ src/recipes/allRecipes.js
@@ -1,15 +1,22 @@
 import React from 'react'
-import { db } from '../firebase'
-import { Link } from 'react-router-dom'
+import "./allrecipes.css"
+// import { db } from '../firebase'
+// import { Link } from 'react-router-dom'
 
 const AllRecipes = () => {
   return (
-    <section className="recipe-add-section">
-      <div className="main-section">
-        <div className="section-head">
-          <h1>ALL RECIPES</h1>
-          <div className="nav-btn-holder">
-            <Link to='/recipes'> <button className="btn-navigate">CREATE A RECIPE</button></Link>
+    <section className="all-recipes">
+      <div className="filter">
+
+      </div>
+      <div className="recipes-body-section">
+        <div className="search-bar-container">
+          <div className="search-bar">
+            <i class="fa-solid fa-magnifying-glass"></i>
+            <input type="text" className='input-field' />
+          </div>
+          <div className="filter-container">
+            <button className="btn btn-add">FILTER <i className="fa-solid fa-filter"></i></button>
           </div>
         </div>
       </div>
